+++ client/package.json
@@ -17,8 +17,7 @@
 		"update-vscode": "vscode-install"
 	},
 	"dependencies": {
-		"vscode-languageclient": "^9.0.1",
-		"@abaplint/core": "^2.113.42"
+		"vscode-languageclient": "^9.0.1"
 	},
 	"devDependencies": {
 		"@types/vscode": "^1.91.0",
